+++ packages/node-core/src/indexer/storeCache/cacheable.ts
@@ -8,9 +8,9 @@ export abstract class Cacheable {
   protected mutex = new Mutex();
 
   abstract clear(blockHeight?: number): void;
-  protected abstract runFlush(tx: Transaction, blockHeight?: number): Promise<void>;
+  protected abstract runFlush(tx: Transaction, blockHeight: number): Promise<void>;
 
-  async flush(tx: Transaction, blockHeight?: number): Promise<void> {
+  async flush(tx: Transaction, blockHeight: number): Promise<void> {
     const release = await this.mutex.acquire();
 
     try {