+++ packages/common/CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Changed
+- Improve error message when project manfiest is invalid (#2408)
 
 ## [3.5.1] - 2024-04-24
 ### Added